requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.2
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
AutoChromedriver==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
browser-cookie3==0.19.1
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
cloudscraper==1.2.71
colorama==0.4.6
Django==4.2.3
djangorestframework==3.14.0
dnspython==2.4.1
exceptiongroup==1.1.2
fastapi==0.100.0
ffmpy==0.3.1
Flask==2.3.2
frozenlist==1.4.0
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lz4==4.3.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mashumaro==3.8.1
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.1
outcome==1.2.0
packaging==23.1
pandas==2.0.3
playwright==1.30.0
pluggy==1.2.0
protobuf==4.23.4
protobuf3-to-dict==0.1.5
psycopg==3.1.9
psycopg2==2.9.6
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.0.3
pydantic_core==2.3.0
pyee==9.0.4
Pygments==2.15.1
pyktok==0.0.16
pymongo==4.4.1
pyparsing==3.1.0
PySocks==1.7.1
pytest==7.4.0
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gql==0.2.4
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
selenium==4.10.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
starlette-graphql==0.2.1
strawberry-graphql==0.195.3
text-unidecode==1.3
tiktok-downloader==0.3.5
tiktok-scraper==1.0.3
TikTokApi==5.2.2
TikTokLive==5.0.8
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.23.1
webdriver-manager==3.8.6
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
